API-4: make sure that only the owner can delete

+++ app/Http/Controllers/Question/DeleteController.php
@@ -12,6 +12,8 @@ class DeleteController extends Controller
      */
     public function __invoke(Question $question)
     {
+        $this->authorize('forceDelete', $question);
+
         $question->forceDelete();
 
         return response()->noContent();
